fix(playground): solve typescript error related to quark's `css` prop
The theme file needs to be in the root folder and not inside the `src`
folder. I still have not figured out why that is necessary.

Using `"include": ["./src/**/*", "./src/*.d.ts"],` did not work. I
assume that the declare module path inside the file does not get picked
up.