chore: handle rollup warning
Rollup warns about not providing an globals for external modules for
reasons explained here: https://github.com/rollup/rollup/issues/1169#issuecomment-268815735

In https://github.com/developit/microbundle/issues/223 the suggestion is
to mark the modules as globals to tell rollup that it's ok.

It seems to me that microbundle should also apply globals for scoped
packages as it does for non-scoped.